refactor(codec): merge Python codec into bridge core
